Read Books From the file and add to the DB- newxt step orders and customers

+++ SQL_Deneme/SQL_Deneme/Book.cs
@@ -20,7 +20,7 @@ namespace SQL_Deneme
 
         public string Genre;
         //public double type;
-        public float Rating;
+        public double Rating;
         public double Price;
         public int Sold;
